f2fs/rapid_gc: Do not wait until all invalid blocks are discarded
Undiscarded blocks may not reach 0 for various reasons, such as
discard granularity settings and discard thread terminating before GC ends,
and when that happens, GC thread enters an indefinite loop.

Remove this logic entirely.

This is not much of an issue since discard thread does not rely on rapid GC.

Change-Id: I82f8c4d2bd3edd06a08d813cbb2832d988cbbd37
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: Dhruv <dhruvgera61@gmail.com>