arch/arm64: dts: disable dfps in jdi fhd panel
Qcom's dfps messes with vsync timers in the current surfaceflinger
implementation and causes micro stutters sometimes. The severity
of stutters is inconsistent across multiple screen wake-ups.

The jdi fhd panel only has two levels of refresh rates (55 and 60).
The ability of switching between 55 and 60 Hz is not worth the
micro stutters. Change it to always operate at 60 Hz.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: Dhruv <dhruvgera61@gmail.com>